+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Next Version
 ### Improvements
  * Added Minecraft version selector to project creation dialog.
+ * Improved appearance of project creation dialog.
 
 
 ## Version 0.5.0-alpha